refactor(DayManager): Day와 DayManager 책임 분리
